backend/requirements.txt
requests
django < 3
ipython
psycopg2-binary
factory_boy
django-mptt
celery
pyyaml
pika
djangorestframework
django-rest-swagger
qbittorrent-api
markdown
django-filter
